+++ lib/testScenario/variableEnv.ts
@@ -85,7 +85,9 @@ export class VariableEnv {
       const variableName = matchResult[1];
       const value = this.get(variableName);
       if (value === undefined) {
-        throw new Error(`Variable not defined while resolving ${source}: ${match}`);
+        // if value not define, just return source name.
+        return source;
+        // throw new Error(`Variable not defined while resolving ${source}: ${match}`);
       }
 
       str = str.substr(0, matchResult.index) + value + str.substr(matchResult.index + match.length);