+++ ChangeLog.md
@@ -1,13 +1,17 @@
 # Change Log - oav
 
+
 ## 08/17/2021 2.8.0
 
 - Isolate RP when running transform on cache model in live validator so that the transform won't fail on global cache level
 - Fix type error in nullableTransformer
 - Throw exception when getting error in transformLoadedSpecs at livevalidator initialization
 - Improve wording of error messages
+- Support resolve body variables in armTemplate
+- Do not write default variable with placeholder
 - Fix validate-traffic file path resolved error.
 - Fix generate-examples can't generate some http code in response
+
 
 ## 08/12/2021 2.7.3